Revert "collections actually work"
This reverts commit 9174f9728e869afef53f869ba3e561ffad1fe7ae.
